backend/requirements.txt
Flask==3.0.0
Flask-MySQLdb==1.0.1
google-generativeai==0.3.0
Pillow==10.0.0
Werkzeug==3.0.0
